refactor(core): extract Project Cut exit owner
