+++ framework/maintainability/abstraction-integrity-report.json
@@ -6,14 +6,14 @@
   "measurement": {
     "aggregates": {
       "kungfu-cli": 20273,
-      "kungfu-flat-root": 36578,
+      "kungfu-flat-root": 36612,
       "kungfu-runtime-family": 6160
     },
     "counts": {
-      "productionFiles": 229,
-      "productionPhysicalLines": 83275,
+      "productionFiles": 230,
+      "productionPhysicalLines": 83309,
       "testFiles": 96,
-      "testPhysicalLines": 45506
+      "testPhysicalLines": 45514
     },
     "hiddenProductionFiles": [],
     "modulesOver1000": [
@@ -67,7 +67,7 @@
       },
       {
         "path": "framework/core/src/python/kungfu/exit_bundle.py",
-        "physicalLines": 2030
+        "physicalLines": 1457
       },
       {
         "path": "framework/core/src/python/kungfu/peer_lifecycle.py",
@@ -114,10 +114,6 @@
       {
         "path": "framework/core/src/python/kungfu/distribution_update.py",
         "physicalLines": 3107
-      },
-      {
-        "path": "framework/core/src/python/kungfu/exit_bundle.py",
-        "physicalLines": 2030
       },
       {
         "path": "framework/core/src/python/kungfu/profile_sdk.py",
@@ -206,10 +202,10 @@
         "responsibilities": 72
       },
       {
-        "contentRoot": "sha256:05a562b0c9508820fa8fc1328a409bc300388b7e7bb975bcc5555d779332a000",
+        "contentRoot": "sha256:5ed800177af8ff6a7b00434b3ba1f6330baf66c27901e91578d2442df9a9ee71",
         "path": "framework/core/src/python/kungfu/exit_bundle.py",
-        "physicalLines": 2030,
-        "responsibilities": 54
+        "physicalLines": 1457,
+        "responsibilities": 36
       },
       {
         "contentRoot": "sha256:216529e0c2108a9ebd74c620a7008c5905e619c9c3aca8ed5677408b5f72cbea",
@@ -370,7 +366,7 @@
       }
     ],
     "schema": "kungfu.python-structure-measurement/v1",
-    "sourceRoot": "sha256:db447763dbdbc750c04391f0d3d2a4db2f2b8ca7eb41e575b7d1ffaf94afb3b5",
+    "sourceRoot": "sha256:4e165d1514c182d1cd4083037b093a9535d17eaee1e3b3ca436c07d6a863d76c",
     "sourceRoots": {
       "production": [
         "framework/core/src/python"
@@ -401,6 +397,12 @@
       "owner": "core/python-release-history",
       "paths": [
         "framework/core/src/python/kungfu/product_release_history.py"
+      ]
+    },
+    {
+      "owner": "core/python-project-cut-exit",
+      "paths": [
+        "framework/core/src/python/kungfu/project_cut_exit.py"
       ]
     },
     {
@@ -441,19 +443,22 @@
   ],
   "publicImports": [
     "kungfu.exit_bundle.ProductReleaseHistoryError",
+    "kungfu.exit_bundle.ProjectCutExitError",
+    "kungfu.project_cut_exit.ProjectCutExitError",
     "kungfu.runtime_broker.RuntimeLeaseManager",
     "kungfu.runtime_broker.RuntimeLifecycleError",
     "kungfu.runtime_broker.NativeReadinessAuthority",
     "kungfu.runtime_service.ProcessRuntimeHost"
   ],
-  "reportRoot": "sha256:482bd76d015d62ffc35c0b96ece023607d485daee123824191e7729d05115d7b",
+  "reportRoot": "sha256:00162ce674ec82744cfb92949ec6b03c0b499002446727303d914374aa33e94d",
   "schema": "kungfu.abstraction-integrity-report/v1",
-  "sourceRoot": "sha256:db447763dbdbc750c04391f0d3d2a4db2f2b8ca7eb41e575b7d1ffaf94afb3b5",
+  "sourceRoot": "sha256:4e165d1514c182d1cd4083037b093a9535d17eaee1e3b3ca436c07d6a863d76c",
   "summary": {
     "blockingIssues": 0
   },
   "typedSeams": [
     "kungfu.product_release_history.ProductReleaseHistoryPort",
+    "kungfu.project_cut_exit.ProjectCutExitPort",
     "kungfu.assignment_lifecycle.ports.AssignmentEventSink",
     "kungfu.assignment_lifecycle.ports.AssignmentRuntime",
     "kungfu.assignment_lifecycle.ports.AssignmentRuntimePort",
